bug-ja/3.6: upgraded to bzr/7016
git-svn-id: https://ssl.jpmoz.net/svn/bug-ja/bugzilla-l10n-ja-3.6/trunk@545 3fd43d0e-4793-11dd-8047-fd5b7de829c9

+++ Bugzilla/Product.pm
@@ -411,7 +411,9 @@ sub remove_from_db {
                                      WHERE series_categories.name = ?',
                                     undef, $self->name);
 
-        $dbh->do('DELETE FROM series WHERE ' . $dbh->sql_in('series_id', $series_ids));
+        if (scalar @$series_ids) {
+            $dbh->do('DELETE FROM series WHERE ' . $dbh->sql_in('series_id', $series_ids));
+        }
 
         # If no subcategory uses this product name, completely purge it.
         my $in_use =
@@ -909,6 +911,17 @@ sub check_product {
     unless ($product) {
         ThrowUserError('product_doesnt_exist',
                        {'product' => $product_name});
+    }
+    return $product;
+}
+
+sub check {
+    my ($class, $params) = @_;
+    $params = { name => $params } if !ref $params;
+    $params->{_error} = 'product_access_denied';
+    my $product = $class->SUPER::check($params);
+    if (!Bugzilla->user->can_access_product($product)) {
+        ThrowUserError('product_access_denied', $params);
     }
     return $product;
 }
